pyproject.toml
[project]
name = "contains_string"
version = "0.0.0"
description = "A validator that validates if an LLM-generated output contains a substring."
authors = [
    {name = "Guardrails AI", email = "contact@guardrailsai.com"}
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">= 3.8.1"
dependencies = [
    "guardrails-ai>=0.4.0"
]

[project.optional-dependencies]
dev = [
    "pyright",
    "pytest",
    "ruff"
] 

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rP"
testpaths = [
    "tests"
]

[tool.pyright]
include = ["validator"]